+++ src/backend/libc/futex/mod.rs
@@ -1,2 +1,2 @@
-#[cfg(not(windows))]
 pub(crate) mod syscalls;
+pub(crate) mod types;
