+++ src/backend/linux_raw/futex/types.rs
@@ -1,10 +1,10 @@
 bitflags::bitflags! {
-    /// `FUTEX_*` flags for use with [`futex`].
+    /// `FUTEX_*` flags for use with the functions in [`futex`].
     ///
-    /// [`futex`]: mod@crate::thread::futex
+    /// [`futex`]: crate::futex
     #[repr(transparent)]
     #[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
-    pub struct FutexFlags: u32 {
+    pub struct Flags: u32 {
         /// `FUTEX_PRIVATE_FLAG`
         const PRIVATE = linux_raw_sys::general::FUTEX_PRIVATE_FLAG;
         /// `FUTEX_CLOCK_REALTIME`
@@ -16,12 +16,12 @@ bitflags::bitflags! {
     }
 }
 
-/// `FUTEX_*` operations for use with [`futex`].
+/// `FUTEX_*` operations for use with the functions in [`futex`].
 ///
-/// [`futex`]: mod@crate::thread::futex
+/// [`futex`]: crate::futex
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 #[repr(u32)]
-pub enum FutexOperation {
+pub enum Operation {
     /// `FUTEX_WAIT`
     Wait = linux_raw_sys::general::FUTEX_WAIT,
     /// `FUTEX_WAKE`
@@ -53,7 +53,7 @@ pub enum FutexOperation {
 }
 
 /// `FUTEX_WAITERS`
-pub const FUTEX_WAITERS: u32 = linux_raw_sys::general::FUTEX_WAITERS;
+pub const WAITERS: u32 = linux_raw_sys::general::FUTEX_WAITERS;
 
 /// `FUTEX_OWNER_DIED`
-pub const FUTEX_OWNER_DIED: u32 = linux_raw_sys::general::FUTEX_OWNER_DIED;
+pub const OWNER_DIED: u32 = linux_raw_sys::general::FUTEX_OWNER_DIED;
