+++ src/backend/linux_raw/thread/mod.rs
@@ -1,2 +1,1 @@
-pub(crate) mod futex;
 pub(crate) mod syscalls;
