+++ tests/futex/basic.rs
@@ -1,22 +1,20 @@
 use core::sync::atomic::{AtomicU32, Ordering};
-use rustix::{
-    io::Errno,
-    thread::{futex, FutexFlags},
-};
+use rustix::futex;
+use rustix::io::Errno;
 
 #[test]
 fn test_lock_unlock_pi() {
     let lock = AtomicU32::new(0);
-    futex::lock_pi(&lock, FutexFlags::empty(), None).unwrap();
+    futex::lock_pi(&lock, futex::Flags::empty(), None).unwrap();
     assert_ne!(lock.load(Ordering::SeqCst), 0);
 
-    let err = unsafe { futex::lock_pi(&lock, FutexFlags::empty(), None).unwrap_err() };
+    let err = futex::lock_pi(&lock, futex::Flags::empty(), None).unwrap_err();
     assert_eq!(err, Errno::DEADLK);
 
-    futex::unlock_pi(&lock, FutexFlags::empty()).unwrap();
+    futex::unlock_pi(&lock, futex::Flags::empty()).unwrap();
     assert_eq!(lock.load(Ordering::SeqCst), 0);
 
-    let err = futex::unlock_pi(&lock, FutexFlags::empty()).unwrap_err();
+    let err = futex::unlock_pi(&lock, futex::Flags::empty()).unwrap_err();
     assert_eq!(err, Errno::PERM);
 }
 
@@ -25,7 +23,7 @@ fn test_lock_unlock_pi() {
 fn test_wait_wake() {
     let lock = std::sync::Arc::new(AtomicU32::new(0));
 
-    match futex::wait(&lock, FutexFlags::empty(), 1, None) {
+    match futex::wait(&lock, futex::Flags::empty(), 1, None) {
         Ok(()) => panic!("Nobody should be waking us!"),
         Err(Errno::AGAIN) => {
             assert_eq!(lock.load(Ordering::SeqCst), 0, "the lock should still be 0")
@@ -38,10 +36,10 @@ fn test_wait_wake() {
         move || {
             std::thread::sleep(std::time::Duration::from_millis(1));
             lock.store(1, Ordering::SeqCst);
-            futex::wake(&lock, FutexFlags::empty(), 1).unwrap();
+            futex::wake(&lock, futex::Flags::empty(), 1).unwrap();
 
             std::thread::sleep(std::time::Duration::from_millis(50));
-            match futex::wait(&lock, FutexFlags::empty(), 1, None) {
+            match futex::wait(&lock, futex::Flags::empty(), 1, None) {
                 Ok(()) => (),
                 Err(Errno::AGAIN) => {
                     assert_eq!(lock.load(Ordering::SeqCst), 2, "the lock should now be 2")
@@ -51,14 +49,14 @@ fn test_wait_wake() {
         }
     });
 
-    match futex::wait(&lock, FutexFlags::empty(), 0, None) {
+    match futex::wait(&lock, futex::Flags::empty(), 0, None) {
         Ok(()) => (),
         Err(Errno::AGAIN) => assert_eq!(lock.load(Ordering::SeqCst), 1, "the lock should now be 1"),
         Err(err) => panic!("{err}"),
     }
 
     lock.store(2, Ordering::SeqCst);
-    futex::wake(&lock, FutexFlags::empty(), 1).unwrap();
+    futex::wake(&lock, futex::Flags::empty(), 1).unwrap();
 
     other.join().unwrap();
 }
@@ -66,13 +64,13 @@ fn test_wait_wake() {
 #[cfg(feature = "std")]
 #[test]
 fn test_timeout() {
-    use rustix::fs::Timespec;
+    use rustix::futex::Timespec;
 
     let lock = AtomicU32::new(0);
 
     let err = futex::wait(
         &lock,
-        FutexFlags::empty(),
+        futex::Flags::empty(),
         0,
         Some(Timespec {
             tv_sec: 1,
@@ -84,7 +82,7 @@ fn test_timeout() {
 
     let err = futex::wait(
         &lock,
-        FutexFlags::empty(),
+        futex::Flags::empty(),
         0,
         Some(Timespec {
             tv_sec: 0,
@@ -96,7 +94,7 @@ fn test_timeout() {
 
     let err = futex::wait(
         &lock,
-        FutexFlags::empty(),
+        futex::Flags::empty(),
         0,
         Some(Timespec {
             tv_sec: -1,
